worker fix

+++ app/workers/ban_post_worker.rb
@@ -4,9 +4,8 @@ class BanPostWorker
   include Sidekiq::Worker
 
   def perform(post_id)
-    byebug
     post = Post.find_by(id: post_id)
-    return if post.blank?
+    return if post.blank? || post.moderated?
 
     post.destroy
   end
